+++ R/rtgstat_package.R
@@ -1,6 +1,5 @@
 #' @import httr2
 #' @import cli
-#' @import retry
 #' @importFrom tidyr replace_na separate unnest_longer unnest_wider
 #' @importFrom dplyr across any_of as_tibble filter left_join mutate pull rename_with select tibble
 #' @importFrom stringr str_glue str_replace_all