+++ R/tg_languages.R
@@ -12,7 +12,8 @@ tg_languages <- function(
   data <- tg_make_request(
     method = 'database/languages',
     token  = tg_get_token(),
-    lang   = lang
+    lang   = lang,
+    check_quote = FALSE
   ) %>%
     tg_parse_response()