+++ R/tg_api_usage.R
@@ -10,28 +10,37 @@ tg_api_usage <- function(){
     check_quote = FALSE
   ) %>%
     tg_parse_response() %>%
-    rename_with(to_snake_case) %>%
-    separate('spent_channels', into = c('spent_channels_fact', 'spent_channels_limit'), sep = '/') %>%
-    separate('spent_requests', into = c('spent_requests_fact', 'spent_requests_limit'), sep = '/') %>%
-    separate('spent_words', into = c('spent_words_fact', 'spent_words_limit'), sep = '/') %>%
+    rename_with(to_snake_case)
+
+  columns_to_flatting <- c('spent_channels', 'spent_words', 'spent_requests')
+
+  for ( col in columns_to_flatting ) {
+
+    col_names <- paste(col, c('fact', 'limit'), sep = '_')
+
+    if ( 'spent_channels' %in% names(data) ) {
+
+      data <-
+        separate(data, .data[[col]], into = col_names, sep = '/') %>%
+        mutate(
+          across(col_names, as.numeric),
+          across(col_names, replace_na, 0)
+        )
+
+    } else {
+
+      data[[col_names[1]]] <- 0
+      data[[col_names[2]]] <- 0
+
+    }
+
+  }
+
+  data <- data %>%
     mutate(expired_at = as.POSIXct(.data$expired_at, origin = '1970-01-01'),
-           across(c('spent_channels_fact',
-                    'spent_channels_limit',
-                    'spent_requests_fact',
-                    'spent_requests_limit',
-                    'spent_words_fact',
-                    'spent_words_limit'),
-                  as.numeric),
-           across(c('spent_channels_fact',
-                    'spent_channels_limit',
-                    'spent_requests_fact',
-                    'spent_requests_limit',
-                    'spent_words_fact',
-                    'spent_words_limit'),
-                  replace_na, 0),
            spent_channels_rate = .data$spent_channels_fact / .data$spent_channels_limit,
            spent_requests_rate = .data$spent_requests_fact / .data$spent_requests_limit,
-           spent_words_rate    = .data$spent_words_fact / .data$spent_words_limit,
+           spent_words_rate    = .data$spent_words_fact    / .data$spent_words_limit,
            across(c('spent_channels_rate',
                     'spent_requests_rate',
                     'spent_words_rate'),