+++ NEWS.md
@@ -1,3 +1,9 @@
+# rtgstat (development version)
+
+* Fixed bug in `tg_set_channel_id()`: `object 'tg_channel_id' not found`.
+* `retry` removed from `rtgstat` dependencies.
+* The mechanism for retrying a request in case of an error has been rewritten to the standard means `httr2`, using `req_retry()`.
+
 # rtgstat 0.1.1
 
 * Fixed errors in DESCRIPTION for CRAN.