+++ NEWS.md
@@ -3,6 +3,7 @@
 * Fixed bug in `tg_set_channel_id()`: `object 'tg_channel_id' not found`.
 * `retry` removed from `rtgstat` dependencies.
 * The mechanism for retrying a request in case of an error has been rewritten to the standard means `httr2`, using `req_retry()`.
+* Add user agent to request in `tg_make_request()` by `req_user_agent()`.
 
 # rtgstat 0.1.1