+++ NEWS.md
@@ -4,9 +4,8 @@
 * `tg_channel_avg_posts_reach()` - Getting the average coverage of channel publications over time.
 * `tg_channel_err()` - Obtaining an ERR indicator for a channel in dynamics.
 
-## Bug fixed
+## Bug fixes
 * Fixed bug in `tg_set_channel_id()`: `object 'tg_channel_id' not found`.
-* `retry` removed from `rtgstat` dependencies.
 
 ## Other
 * `rtgstat` added to [TGStat API documentation](https://api.tgstat.ru/docs/ru/client-libs.html#%D1%8F%D0%B7%D1%8B%D0%BA-r).
@@ -14,6 +13,7 @@
 * Add user agent to request in `tg_make_request()` by `req_user_agent()`.
 * Now API endpoint add to URL by `req_url_path_append()` `inside tg_make_request()`.
 * Disabled checking the api quota usage in the `tg_categories()`, `tg_countries()`, `tg_languages()`.
+* `retry` removed from `rtgstat` dependencies.
 
 # rtgstat 0.1.1